chore: move primitives into separate module
